working branch

+++ src/main/java/model/CartRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CartRepository extends CrudRepository<Cart, Integer> {
+public interface CartRepository extends CrudRepository<Cart, Long> {
     Cart findByCustomer(Customer customer);
     Cart findById(long id);
